separating general into a partial
